Added searching; moved custom iterator
Removed the files "iterator.h" and "list_iterator.h", replacing the iterator for linked_list<T> with one in a nested class. Since all operators needed to be overloaded anyway, this was easier.
Further, a few basic search functions were added (linear search and binary search). The binary search function requires a container with a random access iterator because it tests for inequality.