Fixed binary search
Previously, if an element was outside of the vector on the left side, the program would crash as r would eventually be set to m - 1, which is before vector.begin(). The check has been modified to correct this so the loop never tries to set an iterator before the beginning of the vector.
Also includes minor comment changes.