Added at(), checks for key duplication in insert()
The at() function has been implemented. Throws an exception if the is not found, unlike operator[] (which creates a new element). Also added a check in the insert function that checks for a duplicate key in the linked list before appending. If a duplicate is found, throws a runtime_error